image: fix artefact due to floating blitting precision

+++ pymt/core/image/__init__.py
@@ -186,7 +186,8 @@ class Image(pymt.BaseObject):
 
     def draw(self):
         '''Draw the image on screen'''
-        imgpos = (self.x - self.anchor_x * self.scale, self.y - self.anchor_y * self.scale)
+        imgpos = (int(self.x - self.anchor_x * self.scale),
+                  int(self.y - self.anchor_y * self.scale))
         r, g, b = self.color[:3]
         with DO(gx_color(r, g, b, self.opacity), gx_blending):
             drawTexturedRectangle(texture=self.texture, pos=imgpos, size=(self.size[0] * self.scale, self.size[1] * self.scale))
